requirements.txt
# this file must contain a mirror of all requirements as seen in setup.py!
# (excepted the ones loaded via -e from github or similar, these are contained in the main project's requirements.txt)
Django>=2.1.5,<2.2
MarkupSafe==1.1
Pillow==6.2.0
Celery==4.2.0

beautifulsoup4==4.8.1
django-allauth==0.42.0
django-annoying==0.7.6
django-appconf==1.0.3
django-bigbluebutton==0.1.0
django-bootstrap3-datetimepicker-3==2.6.0
django-bootstrap3==11.0.0
django-classy-tags==0.8.0
django-cors-middleware==1.3.1
django-cron==0.5.0
django-embed-video==0.6
django-el-pagination==2.1.2
django-extra-views==0.6.3
django-filter==2.0.0
django-haystack==2.8.1
django-honeypot==0.6.0
django-jalali==4.0.0
django-mailbox==4.7.1
django-model-utils==1.5.0
django-mptt==0.8.7
django-oauth-toolkit==1.3.2
django-otp==0.5.2
django-phonenumber-field==1.1.0
django-polymorphic==0.7.2
django-sekizai==0.10.0
django-simple-captcha==0.5.7
django-suit==0.2.26
django-taggit==0.23.0
django-widget-tweaks==1.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
dnspython==1.15.0
drf-yasg==1.17.1
easy-thumbnails==2.7
ecdsa==0.13.3
geopy==1.11.0
jsonfield==2.0.2
lxml==4.5.0
html5lib
html2text==2016.9.19
numpy==1.14.5
oauthlib==3.0.1
paramiko==2.4.2
pycrypto==2.6.1
PyJWT==1.7.1
redis==2.10.6
requests==2.20.0
requests-oauthlib==0.8.0
rocketchat-API==0.6.26
six==1.12.0
sqlparse==0.1.10
###############wsgiref==0.1.2
python-dateutil==2.4.1
urllib3==1.24.2
pytz==2018.5
python-memcached==1.59
qrcode==6.1
Unidecode==0.4.21

wagtail==2.3
django-compressor


-e git+git://github.com/wechange-eg/django-awesome-avatar.git@django2#egg=django-awesome-avatar
-e git+git://github.com/wechange-eg/django-filer.git@django2#egg=django-filer
-e git+git://github.com/wechange-eg/django-multiform.git@master#egg=django-multiform
-e git+git://github.com/saschan/django-djajax.git@django2#egg=django-djajax
-e git+git://github.com/wechange-eg/django-select2.git@django2#egg=django-select2
-e git+git://github.com/wechange-eg/django-osm-field.git@django2#egg=django-osm-field
-e git+git://github.com/wechange-eg/python-markdown2.git@master#egg=markdown2
-e git+git://github.com/wechange-eg/pydkim.git@master#egg=pydkim

